hhhhhh
hhhhhhh

+++ cmds/requestrole.js
@@ -2,24 +2,49 @@ const Discord = module.require("discord.js");
 
 module.exports.run = async (bot, message, args) => {
 
+	if (message.channel.id !== "728768539938259005") {
+		return;
+	}
+
 	message.delete();
 	
-	//checking if a member existings in a guild
-	function checkGuild(gid) {
+	function sendMessage(content, time) {
+		message.channel.send(content)
+		.then(msg => {
+			msg.delete(time)
+		})
+	}
+	
+	function checkGuild(gid, sname, rid) {
+		let guild = bot.guilds.get(gid);
+
+		if (guild.member(message.author.id)) {
+			let role = message.guild.roles.get(rid);
+			message.member.addRole(role)
+			sendMessage(":white_check_mark: Found you in **" + sname + "**!", 5000);
+		}
+	}
+		
+	function checkGuild2(gid) {
 		let guild = bot.guilds.get(gid);
 
 		if (guild.member(message.author.id)) {
 			return true;
 		}
 	}
-		
+	
+	checkGuild("614876145250205706", "SmartEye", "728993343790645293");
+	
+	if (!checkGuild2("614876145250205706")) {
+		sendMessage(":x: Couldn't assign you any supported roles!", 7500);
+	} else {
+		sendMessage(":ballot_box_with_check: Properly assigned you to all supported roles!", 7500);
+	}
 	/*if (checkGuild("728756362665918476")) {
 		message.channel.send(":white_check_mark: Found you in Sky.cc!");
 		let role = message.guild.roles.find(role => role.name === "Sky.cc");
 		message.member.addRole(role);
 	}*/
-	
-	//add checkGuild2, pass through guild name and role id on the check to save time?
 }
 
 module.exports.help = {
